fix(ci): python-full-coverage - split oversized files and cover companion schema branches
