+++ tests/unit/node/test_provisioner_parts/test_provisioner_part_002.py
@@ -16,17 +16,19 @@ from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.db.enums import WorkspaceStatus
 from awf.db.repositories import (
+    EgressAuditRepository,
     SecretLeaseRepository,
     WorkspaceRepository,
 )
 from awf.db.session import make_session_factory
 from awf.node.compose_manager import ComposeOperationError, ComposeProjectPaths
-from awf.node.git_manager import GitManager, WorktreeLayout
+from awf.node.egress_policy import LocalEgressPolicyError
+from awf.node.git_manager import GitManager, GitOperationError, WorktreeLayout
 from awf.node.provisioner import (
     Provisioner,
     ProvisionerConfig,
 )
-from awf.profiles.models import ProfileSecret, WorkspaceProfile
+from awf.profiles.models import EgressMode, ProfileSecret, WorkspaceProfile
 from tests.postgres import postgres_test_engine
 
 
@@ -571,3 +573,130 @@ class TestIdempotency:
             reloaded = await WorkspaceRepository(s).get(ws_id)
             assert reloaded is not None
             assert reloaded.status == WorkspaceStatus.ready.value
+
+
+class TestFailureHandlingEdges:
+    @pytest.mark.unit
+    async def test_local_egress_policy_error_marks_workspace_failed(
+        self,
+        session_factory: async_sessionmaker[AsyncSession],
+        git_manager: GitManager,
+        origin_repo: Path,
+    ) -> None:
+        class _EgressFailingLauncher:
+            async def launch(self, _request: Any) -> object:
+                raise LocalEgressPolicyError(
+                    reason_code="LOCAL_EGRESS_MODE_UNSUPPORTED",
+                    mode=EgressMode.restricted,
+                    message="local backend cannot enforce this mode",
+                    details={"network_posture": "restricted"},
+                )
+
+        provisioner = Provisioner(
+            session_factory=session_factory,
+            git=git_manager,
+            stack_launcher=_EgressFailingLauncher(),  # type: ignore[arg-type]
+            config=ProvisionerConfig(node_id="test-node-01"),
+        )
+        async with session_factory() as s:
+            ws = await WorkspaceRepository(s).create(
+                repo_url=str(origin_repo),
+                branch_base="development",
+                task_title="t",
+                task_prompt="p",
+                agent="codex",
+                test_commands=[],
+                resolved_profile={"name": "restricted-local"},
+            )
+            await s.commit()
+            ws_id = ws.id
+
+        with pytest.raises(LocalEgressPolicyError):
+            await provisioner.provision(ws_id)
+
+        async with session_factory() as s:
+            reloaded = await WorkspaceRepository(s).get(ws_id)
+            audit = await EgressAuditRepository(s).get_latest_for_workspace(ws_id)
+            assert reloaded is not None
+            assert reloaded.status == WorkspaceStatus.failed.value
+            assert reloaded.failure_reason == "policy_failure"
+            assert reloaded.failure_message is not None
+            assert "LOCAL_EGRESS_MODE_UNSUPPORTED" in reloaded.failure_message
+            assert audit is None
+            failed_events = [
+                event
+                for event in reloaded.events
+                if event.event_type == "workspace.state_changed"
+                and event.new_state == WorkspaceStatus.failed.value
+            ]
+            assert failed_events[-1].reason_code == "LOCAL_EGRESS_MODE_UNSUPPORTED"
+
+    @pytest.mark.unit
+    async def test_missing_base_branch_marks_workspace_failed(
+        self,
+        provisioner: Provisioner,
+        session_factory: async_sessionmaker[AsyncSession],
+        origin_repo: Path,
+    ) -> None:
+        async with session_factory() as s:
+            ws = await WorkspaceRepository(s).create(
+                repo_url=str(origin_repo),
+                branch_base="nonexistent",
+                task_title="t",
+                task_prompt="p",
+                agent="codex",
+                test_commands=[],
+            )
+            await s.commit()
+            ws_id = ws.id
+
+        with pytest.raises(GitOperationError):
+            await provisioner.provision(ws_id)
+
+        async with session_factory() as s:
+            reloaded = await WorkspaceRepository(s).get(ws_id)
+            assert reloaded is not None
+            assert reloaded.status == WorkspaceStatus.failed.value
+            assert reloaded.failure_reason == "infrastructure_failure"
+            assert reloaded.failure_message is not None
+
+    @pytest.mark.unit
+    async def test_unexpected_provisioning_failure_marks_workspace_failed(
+        self,
+        session_factory: async_sessionmaker[AsyncSession],
+        git_manager: GitManager,
+        origin_repo: Path,
+    ) -> None:
+        class _ExplodingStackLauncher:
+            async def launch(self, request: Any) -> object:
+                raise RuntimeError("template workspace.base.yml.j2 was not found")
+
+        provisioner = Provisioner(
+            session_factory=session_factory,
+            git=git_manager,
+            stack_launcher=_ExplodingStackLauncher(),
+            config=ProvisionerConfig(node_id="test-node-01"),
+        )
+        async with session_factory() as s:
+            ws = await WorkspaceRepository(s).create(
+                repo_url=str(origin_repo),
+                branch_base="development",
+                task_title="t",
+                task_prompt="p",
+                agent="codex",
+                test_commands=[],
+            )
+            await s.commit()
+            ws_id = ws.id
+
+        with pytest.raises(RuntimeError, match="workspace.base.yml.j2"):
+            await provisioner.provision(ws_id)
+
+        async with session_factory() as s:
+            reloaded = await WorkspaceRepository(s).get(ws_id)
+            assert reloaded is not None
+            assert reloaded.status == WorkspaceStatus.failed.value
+            assert reloaded.failure_reason == "infrastructure_failure"
+            assert reloaded.failure_message is not None
+            assert "unexpected provisioning failure" in reloaded.failure_message
+            assert "workspace.base.yml.j2" in reloaded.failure_message
