+++ tests/unit/service/test_workspaces_observability_parts/test_workspaces_observability_part_001.py
@@ -1460,54 +1460,3 @@ def test_recovery_summary_uses_inactive_operator_recovery_operation() -> None:
         "recovery_mode": "validate_only",
     }
     assert "validate-only recovery" in summary.summary
-
-
-@pytest.mark.unit
-def test_recovery_summary_bounds_json_payload_from_previous_recovery_event() -> None:
-    base = datetime(2026, 4, 27, 21, 42, tzinfo=UTC)
-    reverse_at = base + timedelta(seconds=20)
-    payload = {
-        "reason_code": "PAYLOAD_RECOVERY",
-        "action": "retry",
-        "when": reverse_at,
-        "nested": {f"k{index}": index for index in range(33)},
-        "items": list(range(25)),
-        "deep": {"a": {"b": {"c": {"d": {"too": "deep"}}}}},
-        "path": Path("artifact.txt"),
-        **{f"extra_{index}": index for index in range(40)},
-    }
-    workspace = _workspace_for_recovery(
-        created_at=base,
-        events=[
-            _recovery_event(
-                event_id="evt_previous_dispatch",
-                event_type="monitor.recovery_dispatched",
-                occurred_at=base + timedelta(seconds=5),
-                reason_code="RECOVERY_DISPATCH",
-                payload=payload,
-            ),
-            _recovery_event(
-                event_id="evt_reverse",
-                event_type="workspace.state_changed",
-                occurred_at=reverse_at,
-                old_state=WorkspaceStatus.monitoring_pr.value,
-                new_state=WorkspaceStatus.ready.value,
-                reason_code="RECOVERY_DISPATCH",
-            ),
-        ],
-    )
-
-    summary = workspace_recovery_summary(workspace)  # type: ignore[arg-type]
-
-    assert summary is not None
-    assert summary.reason_code == "PAYLOAD_RECOVERY"
-    assert summary.action == "retry"
-    assert summary.recovery_mode is None
-    assert "AWF dispatched retry." in summary.summary
-    assert summary.payload is not None
-    assert summary.payload["when"] == reverse_at.isoformat()
-    assert summary.payload["nested"]["__truncated__"] is True
-    assert summary.payload["items"][-1] == "__truncated__"
-    assert summary.payload["deep"]["a"]["b"]["c"]["d"].startswith("{'too':")
-    assert summary.payload["path"] == "artifact.txt"
-    assert summary.payload["__truncated__"] is True
